poky: sumo refresh 3b8dc3a88e..36d5cee56b
Update poky to sumo HEAD.

Andreas Müller (1):
      libsdl2: Fix left rotated display for RaspPi/VC4/GLES2

Anuj Mittal (2):
      qemux86-directdisk: remove mem= parameter
      initramfs-framework/udev: call settle before kill

Armin Kuster (1):
      libcroco: CVE-2017-7961

Awais Belal (1):
      bitbake: toaster: Fix bitbake flexible path error from Bugzilla #12891

Chong Yi Chai (1):
      mkefidisk: fix installation of kernel image

Paulo Neves (1):
      bitbake: fetcher: Fixed remote removal not throwing exception.

Richard Purdie (1):
      bitbake: main: Fix environment handling for UI module imports

Ross Burton (2):
      security_flags: disable static PIE in glibc
      cmake: put cmake.m4 and toolchain file in PN

Sinan Kaya (1):
      gnupg: CVE-2018-9234

+++ poky/bitbake/lib/bb/fetch2/git.py
@@ -354,10 +354,9 @@ class Git(FetchMethod):
             if not self._contains_ref(ud, d, name, ud.clonedir):
                 needupdate = True
         if needupdate:
-            try: 
-                runfetchcmd("%s remote rm origin" % ud.basecmd, d, workdir=ud.clonedir)
-            except bb.fetch2.FetchError:
-                logger.debug(1, "No Origin")
+            output = runfetchcmd("%s remote" % ud.basecmd, d, quiet=True, workdir=ud.clonedir)
+            if "origin" in output:
+              runfetchcmd("%s remote rm origin" % ud.basecmd, d, workdir=ud.clonedir)
 
             runfetchcmd("%s remote add --mirror=fetch origin %s" % (ud.basecmd, repourl), d, workdir=ud.clonedir)
             fetch_cmd = "LANG=C %s fetch -f --prune --progress %s refs/*:refs/*" % (ud.basecmd, repourl)
